Entrega Final JS

+++ js/main.js
@@ -8,18 +8,15 @@ class Servicio {
   }
 }
 
-const paginasWeb = new Servicio(1, "Paginas Web", 20000, "../images/website.jpg");
-const portfolio = new Servicio(2, "Portfolio", 25000, "../images/portfolio.jpg");
-const tiendaEcommerse = new Servicio(3, "Tienda Ecommerse", 40000, "../images//tiendanube.jpg");
-const autoCad = new Servicio(4, "AutoCad Timon", 25000, "../images/timon.png");
-const render = new Servicio(5, "Render 3D", 30000, "../images/render.jpg");
-const calculos = new Servicio(6, "Calculos", 45000, "../images/velero.png");
-
-
-
+const paginasWeb = new Servicio(1, "Paginas Web", 80, "../images/website.jpg");
+const portfolio = new Servicio(2, "Portfolio", 100, "../images/portfolio.jpg");
+const tiendaEcommerse = new Servicio(3, "Tienda Ecommerse", 150, "../images//tiendanube.jpg");
+const autoCad = new Servicio(4, "AutoCad Timon",90, "../images/timon.png");
+const render = new Servicio(5, "Render 3D", 120, "../images/render.jpg");
+const calculos = new Servicio(6, "Calculos", 180, "../images/velero.png");
 const servicios = [paginasWeb, portfolio, tiendaEcommerse, autoCad, render, calculos];
 
-//Creamos el array carrito 
+
 
 let carrito = [];
 
@@ -30,8 +27,8 @@ if(localStorage.getItem("carrito")) {
 
 const contenedorServicios = document.getElementById("contenedorServicios");
 
-
-const mostrarServicios = () => {
+ 
+const mostrarServicios = () => { 
   servicios.forEach((servicio) => {
       const card = document.createElement("div");
       card.classList.add("col-xl-3", "col-md-6", "col-xs-12");
@@ -40,7 +37,7 @@ const mostrarServicios = () => {
               <img src="${servicio.img}" class="card-img-top imgServicios" alt="${servicio.nombre}">
               <div class="card-body">
               <h5 class="card-title"> ${servicio.nombre} </h5>
-              <p class="card-text"> ${servicio.precio} </p>
+              <p class="card-text"> USD ${servicio.precio} </p>
               <button class="btn colorBoton" id="boton${servicio.id}"> Agregar al Carrito </button>
               </div>
           </div>
@@ -49,7 +46,14 @@ const mostrarServicios = () => {
 
       const boton = document.getElementById(`boton${servicio.id}`);
       boton.addEventListener("click", () => {
-          agregarAlCarrito(servicio.id)
+          agregarAlCarrito(servicio.id);
+          Swal.fire({
+            position: 'top-end',
+            icon: 'success',
+            title: 'Agregaste este servicio al carrito',
+            showConfirmButton: false,
+            timer: 1000
+          })
       })
   })
 }
@@ -97,7 +101,25 @@ const mostrarCarrito = () => {
       //Eliminar productos del carrito: 
       const boton = document.getElementById(`eliminar${servicio.id}`);
       boton.addEventListener("click", () => {
-          eliminarDelCarrito(servicio.id);
+          
+          Swal.fire({
+            title: 'Estas seguro?',
+            text: "Puedes volver a agregar este producto mas tarde si quieres",
+            icon: 'warning',
+            showCancelButton: true,
+            confirmButtonColor: '#3085d6',
+            cancelButtonColor: '#d33',
+            confirmButtonText: 'Si, eliminar!'
+          }).then((result) => {
+            if (result.isConfirmed) {
+              eliminarDelCarrito(servicio.id);
+              Swal.fire(
+                'Eliminado!',
+                'El servicio ha sido elimiado',
+                'success'
+              )
+            }
+          })
       })
   })
   calcularTotal();
@@ -116,13 +138,39 @@ const eliminarDelCarrito = (id) => {
 const vaciarCarrito = document.getElementById("vaciarCarrito");
 
 vaciarCarrito.addEventListener("click", () => {
-  eliminarTodoElCarrito();
+  Swal.fire({
+    title: 'Estas seguro de eliminar los servicios del carrito?',
+    text: "Esta accion no se puede deshacer!",
+    icon: 'warning',
+    showCancelButton: true,
+    confirmButtonColor: '#3085d6',
+    cancelButtonColor: '#d33',
+    confirmButtonText: 'Si, eliminar!'
+  }).then((result) => {
+    if (result.isConfirmed) {
+      eliminarTodoElCarrito();
+      Swal.fire(
+        'Eliminado!',
+        'El carrito ha sido vaciado',
+        'Exitoso!!'
+      )
+    }
+  })
+  
 })
 
 let boton=document.getElementById("finalizarCompra")
 boton.addEventListener("click", respuestaClick)
 function respuestaClick(){
-  alert("Su compra a sido realizada con exito!!!!");
+  Swal.fire({
+    title: 'Gracias por su confianza!!',
+    text: 'Su compra a sido realizada con exito!!!!',
+    imageUrl: '../images/logolp.png',
+    imageWidth: 400,
+    imageHeight: 400,
+    imageAlt: 'Custom image',
+  })
+  
 }
 
 const eliminarTodoElCarrito = () => {
@@ -134,6 +182,20 @@ const eliminarTodoElCarrito = () => {
 }
 
 
+const criptoYa="https://criptoya.com/api/dolar";
+const divDolar = document.getElementById("divDolar");
+setInterval(()=> {
+  fetch(criptoYa)
+  .then(response=>response.json())
+  .then(({solidario})=>{
+    divDolar.innerHTML=`
+    <h4>Tipo de cambio</h4>
+    <p>Dolar solidario: ${solidario}</p>
+    `
+  })
+  .catch(error=> console.error())
+})
+
 const total = document.getElementById("total");
 
 const calcularTotal = () => {
@@ -141,5 +203,5 @@ const calcularTotal = () => {
   carrito.forEach((servicio) => {
       totalCompra += servicio.precio * servicio.cantidad;
   })
-  total.innerHTML = ` $${totalCompra}`;
+  total.innerHTML = ` USD ${totalCompra}`;
 }
